+++ src/hooks/index.js
@@ -1,2 +1,3 @@
 export { default as useSortColumnManager } from './useSortColumnManager';
+export * from './useAuthorityExport';
 export * from './useReportGenerator';